[front] Add Update Room Page
