--- next_app/src/app/admin/create/createRoomPage.module.scss
@@ -1,13 +1,0 @@
-.form {
-  width: 100%;
-  max-width: 500px;
-  margin: auto;
-
-  margin-bottom: 50px;
-}
-
-.button_submit {
-  margin-top: 16px;
-  padding-left: 10px !important;
-  padding-right: 10px !important;
-}
