+++ next_app/src/features/room_admin/utils.ts
@@ -1,0 +1,16 @@
+export const createOptions = (): { value: string; name: string }[] => {
+  return [
+    {
+      value: '',
+      name: 'Select Room Type',
+    },
+    {
+      value: 'video',
+      name: 'Video',
+    },
+    {
+      value: 'manga',
+      name: 'Manga',
+    },
+  ]
+}
